EnemyAi implementation
Enemy first finds the closest barricade and traverses there.
After destroying the barricade or if barricade was already destroyed it finds the nearest turret or player and proceeds to attack that.
After destroying it's target, it tries to find the next target and repeat until no targets remain.
Added probuilder